Make ThinkingIcon it own file
